requirements.txt
certifi
fastapi==0.68.1
langchain==0.0.250
numpy==1.24.3
openai==1.11.1
pandas==2.0.1
python-dotenv==1.0.0
pymongo==4.4.0
regex==2023.6.3
requests==2.30.0
retry==0.9.2
tiktoken==0.4.0
openpyxl==3.1.2
uvicorn==0.12.3